moved config related files to config.py in utils
