+++ utils/log.py
@@ -1,29 +1,4 @@
 import logging
-import logging.config
-import sys
-import os
-
-def logPath() -> str:
-    """
-    Log file is created in the user's home directory.
-
-    Returns:
-        str: The path to the log file.
-    """
-    return os.path.join(dataDir(), "log.txt")
-
-
-def dataDir() -> str:
-    """
-    Data directory is created in the user's home directory.
-
-    Returns:
-        str: The path to the data directory.
-    """
-    directory = os.path.join(os.path.expanduser("~"), ".pictopy")
-    os.makedirs(directory, exist_ok=True)
-    return directory
-
 
 class StreamToLogger(object):
     """
@@ -41,41 +16,3 @@ class StreamToLogger(object):
 
     def flush(self):
         pass  # StreamHandler does not require flush implementation
-
-
-LOG_CONFIG = {
-    "version": 1,
-    "disable_existing_loggers": False,
-    "formatters": {
-        # for console since you dont need that much info on the console
-        'console': {
-            'format': '[%(levelname)s] %(message)s',
-        },
-        "standard": {
-            "format": "[%(levelname)s|%(module)s|L%(lineno)d] %(asctime)s: %(message)s",
-            "datefmt": "%Y-%m-%dT%H:%M:%S%z",
-        },
-    },
-    "handlers": {
-        "default": {
-            "level": "INFO",
-            "formatter": "console",
-            "class": "logging.StreamHandler",
-            "stream": sys.stdout,  # Output to stdout
-        },
-        "file": {
-            "level": "INFO",
-            "formatter": "standard",
-            "class": "logging.handlers.RotatingFileHandler",
-            "filename": logPath(),
-            "maxBytes": 1024 * 1024 * 5,  # 5 MB
-            "backupCount": 5,
-        },
-    },
-    "loggers": {
-        "": {  # Root logger
-            "level": "DEBUG",
-            "handlers": ["default", "file"],
-        },
-    },
-}
